+++ R/stream.R
@@ -25,9 +25,11 @@ stream.mdl_df <- function(object, new_data, ...){
 }
 
 #' @export
-stream.lst_mdl <- function(object, new_data, ...){
+stream.mdl_lst <- function(object, new_data, ...){
   add_class(map2(object, new_data, stream, ...), class(object))
 }
+#' @export
+stream.lst_mdl <- deprecate_lst_mdl(stream.mdl_lst)
 
 #' @export
 stream.mdl_ts <- function(object, new_data, ...){